Unified Alert/requirements.txt
facebook-business==19.0.0
google-cloud-bigquery==3.14.1
google-cloud-secret-manager==2.18.1
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
pandas==2.1.4
pandas-gbq>=0.18.0
numpy==1.26.2
streamlit==1.29.0
plotly==5.18.0
Flask==3.0.0
python-dotenv==1.0.0
google-ads>=23.0.0